automation: align Hudi smoke prep types with FDW schema
dlagent's HudiUtilities.validateTypeCompatible enforces an exact match
between the Hudi schema (read from .hoodie metadata) and the foreign-
table column type.  The prep SQLs and the FDW SQLs disagreed on three
columns, so every read failed before returning a row:

* id: prep emitted Spark INT literals; FDW declared bigint.  Cast the
  literal in the first UNION branch to BIGINT so the entire column
  promotes to long, and use sequence(CAST(1 AS BIGINT), ...) for the
  bulk explode generator.  Also flip the two FDW tables that still
  declared "id int" to bigint.

* price/amount: prep used DECIMAL(10,2); FDW declared double precision.
  Switch the prep to CAST AS DOUBLE - keeps numeric values identical
  in psql output (no trailing zeros to regenerate elsewhere).

* ts_col (hudi_cow_types): prep emitted TIMESTAMP; FDW declared bigint.
  Switch to CAST AS BIGINT so the file matches.

Expected outputs for the three Hudi sub-tests (hudi_smoke,
hudi_coverage, hudi_btree_merger) refresh to reflect the merged-on-
read view that Hudi 0.13.x returns now that schema validation passes
and Cloudberry can actually read the data.

+++ contrib/datalake_fdw/automation/sqlrepo/smoke/hudi/sql/hudi_smoke.sql
@@ -110,7 +110,7 @@ GROUP BY region ORDER BY region;
 -- ============================================================
 DROP FOREIGN TABLE IF EXISTS hudi_null_table;
 CREATE FOREIGN TABLE hudi_null_table (
-    id int,
+    id bigint,
     name text,
     val int,
     ts bigint
